requirements.txt
# wsgi server
gunicorn[gevent]==20.1.0

# Flask
Flask==2.3.2

# common
python-dotenv==0.21.1
cryptography==39.0.1
marshmallow==3.20.1
requests==2.31.0
cachetools==5.3.0

# linting and testing
pytest==7.4.0
mypy==1.5.1
black==23.7.0
types-requests==2.31.0.2
types-cachetools==5.3.0.6
Flask-SQLAlchemy==3.0.5

